Added error handling for JS Partials
Previously, if the partial JS file can't be found, the module would simply spit out an `undefined`, which would cause an error in the JS file. Now, if the partial JS file is missing, we will get a console error message that clearly explains that the required partial JS file is missing.